Update Cargo.lock, Cargo.toml, config.rs, and 7 more files
Implemented data layer utilizing SurrealDB with RocksDB as the file-based implementation. Memory or file-based is determined by the new DatabaseType enumeration value at config.database.databaseType. The config.database.path value is only used when creating an instance of a file-based database, and it is stored alongside the logs in the local data directory.

Added a test in data/dal.rs.