Update MPTest.tscn, Ui.cs, VttClient.cs, and 3 more files
Added UserManager struct on the server, manages association between User ID and dynamically generated Client ID; The server applies a Client ID to each client upon authentication. This Client ID will remain static between reconnects as long as the server is running but will be regenerated upon client connection every time the server restarts.

Updated the client to be capable of sending a username along with the AuthenticateSend command; The server does not yet actually perform authentication but it let me stub out the command flow for Client ID generation.

Updated the client to accept a Client ID from the AuthenticateSuccess command's data.