Update dal.rs, db.rs, user.rs, and 2 more files
Added DataLayer::userFind to look up a user based on their username rather than id.
Simplified the Database struct; It can only exist with a valid Surreal<Db> instance.
Renamed User::Name to User::ResourceName, to be a bit more clear about what it is.

Updated client.rs and server.rs to clone the Config instance around so that client.rs can create an instance of DataLayer; This cannot be the final form of accessing the database as multiple client web socket tasks need to be able to process queries simultaneously.